tp2/requirements.txt
PyOpenGL==3.1.0
numpy==1.17.2
pillow==6.1.0
pynput==1.4.4